Reduce filecount
